requirements.txt
anyio==3.5.0
asgiref==3.5.0
click==8.0.4
colorama==0.4.4
fastapi==0.75.0
h11==0.13.0
idna==3.3
mysql-connector-python==8.0.28
protobuf==3.19.4
pydantic==1.9.0
python-dotenv==0.19.2
sniffio==1.2.0
starlette==0.17.1
typing_extensions==4.1.1
uvicorn==0.17.6
